refactor: 버퍼풀 패키지 생성
