+++ src/test/java/database/innodb/bufferpool/BufferPoolTest.java
@@ -1,4 +1,4 @@
-package database.innodb;
+package database.innodb.bufferpool;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
